EDITME: cover title for meson-dm-usb
# Lines starting with # will be removed from the cover letter. You can use
# them to add notes or reminders to yourself.

EDITME: describe the purpose of this series. The information you put here
will be used by the project maintainer to make a decision whether your
patches should be reviewed, and in what priority order. Please be very
detailed and link to any relevant discussions or sites that the maintainer
can review to better understand your proposed changes. If you only have a
single patch in your series, the contents of the cover letter will be
appended to the "under-the-cut" portion of the patch.

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated during
# the b4 send stage. You can also run "b4 prep --auto-to-cc" to auto-populate
# the To: and Cc: trailers based on the code being modified.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20221024-meson-dm-usb-60e413696519",
    "base-branch": null
  }
}